cc: Use SkMScalar instead of doubles for transforms in cc
BUG=269819

Review URL: https://chromiumcodereview.appspot.com/23043011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@222683 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 803f6b54878c02a34499262aa9a6f9cd8813d1b1

+++ cc/trees/damage_tracker_unittest.cc
@@ -340,14 +340,12 @@ TEST_F(DamageTrackerTest, VerifyDamageForTransformedLayer) {
   // should increase the size of the expected rect by sqrt(2), centered around
   // (100, 100). The old exposed region should be fully contained in the new
   // region.
-  double expected_width = 30.0 * sqrt(2.0);
-  double expected_position = 100.0 - 0.5 * expected_width;
-  gfx::RectF expected_rect(expected_position,
-                           expected_position,
-                           expected_width,
-                           expected_width);
+  float expected_width = 30.f * sqrt(2.f);
+  float expected_position = 100.f - 0.5f * expected_width;
+  gfx::RectF expected_rect(
+      expected_position, expected_position, expected_width, expected_width);
   root_damage_rect =
-          root->render_surface()->damage_tracker()->current_damage_rect();
+      root->render_surface()->damage_tracker()->current_damage_rect();
   EXPECT_FLOAT_RECT_EQ(expected_rect, root_damage_rect);
 }
